Allow user to upload audio for deep clustering
Caveat: At the moment hacked together so model path is hardcoded in and
deep-clustering-extras is required. Specifically a copy of the nussl
repo on the deep-clustering-extras branch is expected to be in the same
directory as the top of the wut directory (as if one was in a directory
and cloned wut and nussl successively). Additionally, there appears to
be some confusion with the current old version of nussl such that to
test this one must rename that directory to avoid naming conflicts (`mv
nussl old_nussl`). These issues will be ironed out before this code is
actually merged into master.