Remove nussl
This branch depends on a more current version of nussl than the version
deleted in this commit. Currently that newer version is required to be
in the same directory containing this repo (i.e. if you've cloned this
into `/Users/shuri/` then you also need to clone nussl into
`/Users/shuri`). This commit deletes the old nussl inside this repo to
avoid import collisions. In the future, nussl will be either pip
installed or set up as a git submodule for usage of experimental nussl
features.